Updating file structure and HISTORY.rst
