+++ HISTORY.rst
@@ -1,10 +1,10 @@
 History
 -------
 
-2.9.0 (2013-05-xx)
+2.9.0 (2013-05-04)
 ++++++++++++++++++
 
-- Fixed streaming issue #144, added ``TwythonStreamer`` and ``TwythonStreamHandler`` to aid users in a friendly streaming experience
+- Fixed streaming issue #144, added ``TwythonStreamer`` to aid users in a friendly streaming experience (streaming examples in ``examples`` and README's have been updated as well)
 
 2.8.0 (2013-04-29)
 ++++++++++++++++++
