+++ twython/streaming/api.py
@@ -1,84 +1,12 @@
-from . import __version__
-from .compat import json
-from .exceptions import TwythonStreamError
+from .. import __version__
+from ..compat import json
+from ..exceptions import TwythonStreamError
+from .types import TwythonStreamerTypes
 
 import requests
 from requests_oauthlib import OAuth1
 
 import time
-
-
-class TwythonStreamStatuses(object):
-    """Class for different statuses endpoints
-
-    Available so TwythonStreamer.statuses.filter() is available.
-    Just a bit cleaner than TwythonStreamer.statuses_filter(),
-    statuses_sample(), etc. all being single methods in TwythonStreamer
-    """
-    def __init__(self, streamer):
-        self.streamer = streamer
-
-    def filter(self, **params):
-        """Stream statuses/filter
-
-        Accepted params found at:
-        https://dev.twitter.com/docs/api/1.1/post/statuses/filter
-        """
-        url = 'https://stream.twitter.com/%s/statuses/filter.json' \
-              % self.streamer.api_version
-        self.streamer._request(url, 'POST', params=params)
-
-    def sample(self, **params):
-        """Stream statuses/sample
-
-        Accepted params found at:
-        https://dev.twitter.com/docs/api/1.1/get/statuses/sample
-        """
-        url = 'https://stream.twitter.com/%s/statuses/sample.json' \
-              % self.streamer.api_version
-        self.streamer._request(url, params=params)
-
-    def firehose(self, **params):
-        """Stream statuses/filter
-
-        Accepted params found at:
-        https://dev.twitter.com/docs/api/1.1/get/statuses/firehose
-        """
-        url = 'https://stream.twitter.com/%s/statuses/firehose.json' \
-              % self.streamer.api_version
-        self.streamer._request(url, params=params)
-
-
-class TwythonStreamTypes(object):
-    """Class for different stream endpoints
-
-    Not all streaming endpoints have nested endpoints.
-    User Streams and Site Streams are single streams with no nested endpoints
-    Status Streams include filter, sample and firehose endpoints
-    """
-    def __init__(self, streamer):
-        self.streamer = streamer
-        self.statuses = TwythonStreamStatuses(streamer)
-
-    def user(self, **params):
-        """Stream user
-
-        Accepted params found at:
-        https://dev.twitter.com/docs/api/1.1/get/user
-        """
-        url = 'https://userstream.twitter.com/%s/user.json' \
-              % self.streamer.api_version
-        self.streamer._request(url, params=params)
-
-    def site(self, **params):
-        """Stream site
-
-        Accepted params found at:
-        https://dev.twitter.com/docs/api/1.1/get/site
-        """
-        url = 'https://sitestream.twitter.com/%s/site.json' \
-              % self.streamer.api_version
-        self.streamer._request(url, params=params)
 
 
 class TwythonStreamer(object):
@@ -122,10 +50,10 @@ class TwythonStreamer(object):
         self.retry_count = retry_count
 
         # Set up type methods
-        StreamTypes = TwythonStreamTypes(self)
+        StreamTypes = TwythonStreamerTypes(self)
         self.statuses = StreamTypes.statuses
-        self.__dict__['user'] = StreamTypes.user
-        self.__dict__['site'] = StreamTypes.site
+        self.user = StreamTypes.user
+        self.site = StreamTypes.site
 
     def _request(self, url, method='GET', params=None):
         """Internal stream request handling"""
